backend/requirements.txt
Flask>=2.0
PyPDF2>=3.0
python-docx>=0.8.11
reportlab>=3.6
Werkzeug>=2.0
